Implement Phase 2: reconciliation engine, DB persistence, MQTT, and HTTP agent protocol
Replace gRPC with HTTP/JSON on the agent side to meet the <5MB binary
target (gRPC added 6.7MB). Agent binary is now 2.6MB on linux/arm64
after UPX compression (down from 12MB in Phase 1).

Agent changes:
- Rewrite main.go to wire BoltDB, HTTP client, MQTT, health collector,
  reconciler plugins, and graceful shutdown (offline-first startup)
- Add exponential backoff retry to HTTP client (3 attempts, 1s/2s/4s)
- Register FileManager and ServiceManager plugins on the reconciler
- Fix filemanager test for cross-platform POSIX mode handling

Controller changes:
- Add AgentApiController with REST endpoints for agent communication
  (register, heartbeat, desired-state, report-state)
- Replace in-memory DeviceRegistry with JPA repository-backed service
- Add Flyway V1 migration (devices + device_manifests tables)
- Add MQTT integration (TelemetryListener, CommandPublisher, MqttConfig)
- Add agent DTO records matching the agent's JSON contract

Build & docs:
- build-agent.sh now applies UPX compression and enforces 5MB size gate
- ADR-001: document HTTP/JSON over gRPC decision with measured evidence
- ADR-002: document UPX compression approach, trade-offs, and rationale

+++ controller/src/main/java/com/edgeguardian/controller/api/DeviceController.java
@@ -47,7 +47,7 @@ public class DeviceController {
     }
 
     @GetMapping("/count")
-    public Mono<Integer> countDevices() {
+    public Mono<Long> countDevices() {
         return Mono.just(registry.count());
     }
 }
